+++ frontend/src/app/(authorized)/settings/form.tsx
@@ -1,6 +1,7 @@
+import { useEffect, useState } from "react";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { useForm } from "react-hook-form";
-import { object, string as yupString, ref } from "yup";
+import { FirebaseError } from "firebase/app";
 import {
   EmailAuthProvider,
   GoogleAuthProvider,
@@ -21,46 +22,7 @@ import {
 } from "@/components/Form/FormField";
 import useFileUploader from "@/hooks/useFileUploader";
 import { UserData } from "@/common/types";
-import { useEffect, useState } from "react";
-import { FirebaseError } from "firebase/app";
-
-const formSchema = object().shape({
-  displayName: yupString()
-    .min(3, "Name is too short")
-    .max(30, "Name is too long")
-    .required("Name is required"),
-  email: yupString().email("Invalid email").required("Email is required"),
-  password: yupString().min(
-    6,
-    "Password length should be at least 6 characters"
-  ),
-  passwordConfirm: yupString()
-    .oneOf([ref("password")], "Passwords must match")
-    .when("password", {
-      is: (password: string) => password?.length > 0,
-      then: (schema) => schema.required("Password confirmation is required"),
-      otherwise: (schema) => schema,
-    }),
-});
-
-const handleFirebaseError = (error: FirebaseError) => {
-  switch (error.code) {
-    case "auth/email-already-in-use":
-      return "Email already in use";
-    case "auth/invalid-email":
-      return "Invalid email";
-    case "auth/weak-password":
-      return "Password is too weak";
-    case "auth/wrong-password":
-      return "Wrong password";
-    case "auth/missing-password":
-      return "Please enter your current password";
-    case "auth/too-many-requests":
-      return "Too many requests. Try again later";
-    default:
-      return "Something went wrong";
-  }
-};
+import { formSchema, handleFirebaseError } from "./schema";
 
 const Form = ({ email, displayName }: Omit<UserData, "photoUrl">) => {
   const [needsReauthentication, setNeedsReauthentication] = useState(false);
@@ -119,7 +81,6 @@ const Form = ({ email, displayName }: Omit<UserData, "photoUrl">) => {
     } catch (error) {
       if (error instanceof FirebaseError) {
         toast.error(handleFirebaseError(error));
-        console.error(error);
       } else {
         toast.error("Something went wrong");
         console.error(error);
